requirements.txt
absl-py==2.4.0
bottle==0.13.4
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.0
fsspec==2026.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
proxy_tools==0.1.0
pycparser==3.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-WebKit==12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyvirtualcam==0.15.0
pywebview==6.1
PyYAML==6.0.3
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing_extensions==4.15.0
